# contracts/openapi/core.yaml
openapi: 3.1.0
info:
  title: Ghost core API
  version: 0.1.0
  description: |
    Core Go service. Public endpoints back the web UI (and later Slack/MCP). The `internal`
    endpoints are the bounded pull-context tools the draft agent calls; every call is logged
    server-side in context_access_log under the run the bearer run token names.
servers:
  - url: http://127.0.0.1:8080
security:
  - apiToken: []
...
paths:
...
  /accounts/{account_id}/graph:
    get:
      tags: [accounts]
      operationId: getAccountGraph
      description: |
        The bounded neighborhood of the account in the Neo4j projection: account, opportunities, active
        stakeholders, recent activities, supported claims and commitments, recent signals, decision
        episodes and applicable knowledge (HAR-96 Architecture update, HAR-129 1B). The web account map and
        the account agent read this same projection (the agent through the `graph_neighborhood` context
        tool). Nodes and edges are ids, short labels, status, validity and evidence refs only: Neo4j
        answers what is connected, Postgres answers exactly what happened (claim values and quotes are
        served by `/accounts/{account_id}/claims` and the timeline).

        `limit` caps the nodes of each section (default 10, max 20) and `truncated` says when a section
        had more. By default only what currently holds is returned; `include_closed=true` adds history
        (closed relationships, superseded claims, expired signals, every knowledge status). The
        operator view is not visibility-filtered (like the other operator reads); the run-scoped tool is.

        `projection.complete` is false while the account has an unfinished projection job: the graph
        then lags Postgres and `projection.projected_at` says when it last caught up.

        `world_as_of` returns the neighborhood as the world stood strictly before that time (ADR-0019):
        the same builder the projector uses, run over Postgres with the cutoff, so the response has the
        same shape and ids but no node or edge derived from an activity at or after the time. Claim
        status, supersession, closed relationships and signal expiry are those at that time. It does not
        read Neo4j: `projection.complete` is true and `projected_at` null, and it works while the graph
        database is down. Knowledge, rejected claims and a person's `title` have no world time and are
        left out.
      parameters:
        - { $ref: "#/components/parameters/AccountId" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 10 }
        - name: include_closed
          in: query
          schema: { type: boolean, default: false }
        - name: world_as_of
          in: query
          description: World-time cutoff, strictly before; see above.
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Bounded neighborhood
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Graph" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /events/{event_id}/graph-diff:
    get:
      tags: [accounts]
      operationId: getEventGraphDiff
      description: |
        The exact nodes and edges the graph projections of one source event added, changed or removed
        (HAR-129 1B "show exact graph diff"). The diff of each projection is recorded in Postgres before
        the graph is written; an event is usually projected twice (ingest, then recompute) and the
        projections are merged into the net change. `attributed_to_event` marks the elements the event is
        itself evidence for; the others changed in the same coalesced burst. `projected: false` means no
        projection of the event has recorded its diff yet.
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
      responses:
        "200":
          description: Graph diff of the event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphDiff" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /graph/projection:
    get:
      tags: [accounts]
      operationId: getGraphProjection
      description: |
        Projection lag: how far the Neo4j graph is behind Postgres (the outbox of unfinished projection
        jobs) and the enqueue-to-complete latency of the last 100 projections. No agent trigger fires for
        an account while its projection is unfinished.
      responses:
        "200":
          description: Projection lag
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphLag" }
        "401": { $ref: "#/components/responses/Error" }
...
  /ask:
    post:
      tags: [ask]
      operationId: askCliff
      description: |
        Ask Cliff (HAR-129 contract change approved by the product owner, 2026-10-06): one free-form question from a
        Slack DM or an @mention thread, answered by a tool-using agent. Transport-neutral: the Slack adapter only
        forwards the text and who asked; everything else happens here. Core mints a short-lived ask token, calls the
        worker's POST /v1/ask, and serves that agent's read tools at POST /internal/ask/tools/{tool} (the existing
        read APIs only: no new data, and the worker never touches the database). The answer comes only from tool
        results, cites them, ends with deep links into the web control plane (built from GHOST_WEB_BASE_URL), and says
        "I don't know" when the tools do not show it. State-changing demo actions (`play_next`, `demo_status`)
        are never executed here: the answer carries a `proposed_action` (its `summary` is written by core from the action's
        kind and the replay, never taken from the model) that the caller must confirm, then
        sends to POST /ask/actions. There is no tool that sends an email or writes the CRM; drafts and CRM previews
        are dry runs, labelled DRAFT. A free-form question is a live paid model call (about 2 to 6 calls) under the
        same spend guard as every other call; it is not part of the recorded demo path.
        502/424 carry the worker's error code; 503 `ask_unavailable` when the deployment has no worker.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AskRequest" }
      responses:
        "200":
          description: The answer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AskAnswer" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "424": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /ask/actions:
    post:
      tags: [ask]
      operationId: runAskAction
      description: |
        Run one confirmed demo action proposed by POST /ask. The adapter calls this only after the human pressed Run.
        `play_next` posts the same core endpoint the web Play control posts
        (`POST /replay/manifests/{manifest_id}/episodes/next`, the manifest of GHOST_DEMO_MANIFEST_ID), so the recorded
        answers replay exactly as they do for web Play. `demo_status` reads the manifest's progress. The replay conflicts
        (`replay_complete`, `play_in_progress`) come back as a refused result with 200, not as an error.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AskActionRequest" }
      responses:
        "200":
          description: The action's result (done or refused, with the words to show)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AskActionResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /internal/ask/tools/{tool}:
    post:
      tags: [internal]
      operationId: runAskTool
      description: |
        One read tool of the Ask Cliff agent, called by the worker with the ask token minted for that question
        (`at1.<expiry unix s>.<nonce>.<hex hmac-sha256>`; its key is domain-separated from the run token's, so neither
        token is accepted as the other, and the operator `apiToken` is not accepted here). Every tool is a read of an
        existing core API served in process (GET only): `list_accounts`, `account_state` (`as_of` is world time),
        `timeline` (`from`, `to`), `graph_neighborhood`, `episode`, `gate_results`, `strategies`, `human_decision`,
        `judgment_inference`, `knowledge`, `knowledge_attribution` (retrieved, applicable and used are three separate
        lists, and nothing here is called influence), `search_activities`, `draft_followup` (dry run, labelled DRAFT)
        and `crm_update_preview` (what would be written; nothing is). `account` accepts an account id or a name. The
        result is bounded (`truncated`), carries the deep `links` a citation may use, and `empty` is true when the data
        shows nothing. An unknown tool is a 404 and bad arguments are a 400.
      security:
        - askToken: []
      parameters:
        - name: tool
          in: path
          required: true
          schema: { type: string, pattern: "^[a-z_]{1,48}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                args: { type: object }
      responses:
        "200":
          description: The tool result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AskToolResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    apiToken:
      type: http
      scheme: bearer
      description: Operator/UI token (GHOST_API_TOKEN). Slack and MCP adapters use their own service tokens.
    runToken:
      type: http
      scheme: bearer
      description: Run-scoped token handed to the worker in /v1/draft; binds context pulls to one run and its account.
    askToken:
      type: http
      scheme: bearer
      description: Question-scoped token handed to the worker in /v1/ask; short-lived, read tools only.
  parameters:
...
    AccountId:
      name: account_id
      in: path
      required: true
      schema: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    AskRequest:
      type: object
      additionalProperties: false
      required: [text, channel_kind, user]
      properties:
        text: { type: string, minLength: 1, maxLength: 2000, description: "The question, with the bot mention already removed." }
        channel_kind:
          type: string
          enum: [dm, thread]
          description: Where the question came from; the adapter replies in the DM, or in a thread under the mention.
        user: { type: string, minLength: 1, maxLength: 64, description: "The asker's Slack member id." }
        thread_ref: { type: string, maxLength: 200, description: "Opaque reference to the conversation (channel:thread_ts), logged only." }
    AskCitation:
      type: object
      additionalProperties: false
      required: [call_id, tool, label]
      properties:
        call_id: { type: string, minLength: 1, maxLength: 16, description: The tool call of this question that returned the cited data. }
        tool: { type: string, minLength: 1, maxLength: 48 }
        label: { type: string, minLength: 1, maxLength: 200 }
        url: { type: [string, "null"], maxLength: 2000, description: "Deep link into the web control plane (episode, gate, trace span, account graph); null when the tool has none." }
    AskProposedAction:
      type: object
      additionalProperties: false
      required: [kind, summary, requires_confirmation]
      properties:
        kind: { type: string, enum: [play_next, demo_status] }
        summary: { type: string, minLength: 1, maxLength: 400, description: "What running it will do, in product words." }
        requires_confirmation: { type: boolean, const: true }
    AskAnswer:
      type: object
      additionalProperties: false
      required: [answer_markdown, citations]
      properties:
        answer_markdown: { type: string, minLength: 1, maxLength: 12000, description: "The answer, ending with deep links into the control plane when the data has any." }
        citations: { type: array, maxItems: 20, items: { $ref: "#/components/schemas/AskCitation" } }
        proposed_action:
          description: A state-changing demo action Cliff offers; nothing has run. The caller must confirm it first.
          $ref: "#/components/schemas/AskProposedAction"
    AskActionRequest:
      type: object
      additionalProperties: false
      required: [kind, user]
      properties:
        kind: { type: string, enum: [play_next, demo_status] }
        user: { type: string, minLength: 1, maxLength: 64, description: The Slack member id of the human who pressed Run. }
    AskActionResult:
      type: object
      additionalProperties: false
      required: [status, message_markdown]
      properties:
        status: { type: string, enum: [done, refused] }
        message_markdown: { type: string, minLength: 1, maxLength: 4000 }
        reason: { type: string, maxLength: 64, description: "Machine reason of a refusal, for example replay_complete, play_in_progress or not_configured." }
    AskToolResult:
      type: object
      additionalProperties: false
      required: [tool, ok, empty, truncated, data, links]
      properties:
        tool: { type: string }
        ok: { type: boolean, description: "False when the arguments named something that does not exist; `data` then holds an `error` text." }
        empty: { type: boolean, description: "True when the tool found nothing, so the data shows nothing." }
        truncated: { type: boolean }
        dry_run: { type: boolean, description: "True on the draft and CRM-preview tools. Nothing was sent or written." }
        data: { description: "What the existing API returned, bounded." }
        links:
          type: array
          maxItems: 12
          items:
            type: object
            additionalProperties: false
            required: [label, url]
            properties:
              label: { type: string }
              url: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
...
    GraphEvidenceRef:
      type: object
      additionalProperties: false
      required: [activity_id]
      properties:
        activity_id: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
    Graph:
      type: object
      required: [account_id, nodes, edges, sections, truncated, withheld, projection]
      properties:
        account_id: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
        nodes:
          type: array
          items:
            type: object
            required: [id, type, label, evidence_refs, source_event_ids]
            properties:
              id: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
              type: { type: string, description: "A node label of contracts/graph/ontology.v1.json." }
              label: { type: string }
              status: { type: string }
              valid_from: { type: string }
              valid_to: { type: string }
              data: { type: object }
              evidence_refs: { type: array, items: { $ref: "#/components/schemas/GraphEvidenceRef" } }
              source_event_ids: { type: array, items: { $ref: "../schemas/common.v1.json#/$defs/uuid" } }
              withheld: { type: string, enum: [visibility] }
        edges:
          type: array
          items:
            type: object
            required: [id, source, target, rel_type, status, evidence_refs, source_event_ids]
            properties:
              id: { type: string }
              source: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
              target: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
              rel_type: { type: string, description: "A relationship type of contracts/graph/ontology.v1.json." }
              status: { type: string }
              standing: { $ref: "../schemas/common.v1.json#/$defs/standing" }
              confidence: { type: number }
              valid_from: { type: string }
              valid_to: { type: string }
              evidence_refs: { type: array, items: { $ref: "#/components/schemas/GraphEvidenceRef" } }
              source_event_ids: { type: array, items: { $ref: "../schemas/common.v1.json#/$defs/uuid" } }
        sections:
          type: object
          description: Node ids per section, in reading order.
          additionalProperties: { type: array, items: { $ref: "../schemas/common.v1.json#/$defs/uuid" } }
        truncated: { type: boolean }
        withheld: { type: boolean }
        projection:
          type: object
          required: [complete, projected_at]
          properties:
            complete: { type: boolean }
            projected_at: { type: [string, "null"] }
    GraphDiff:
      type: object
      required: [event_id, projected, job_ids, summary, changes]
      properties:
        event_id: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
        projected: { type: boolean }
        job_ids: { type: array, items: { type: integer } }
        summary:
          type: object
          required: [added, changed, removed, repaired]
          properties:
            added: { type: integer }
            changed: { type: integer }
            removed: { type: integer }
            repaired: { type: integer }
        changes:
          type: array
          items:
            type: object
            required: [kind, op, type, id, source_event_ids, attributed_to_event]
            properties:
              kind: { type: string, enum: [node, edge] }
              op: { type: string, enum: [added, changed, removed, repaired] }
              type: { type: string }
              id: { type: string }
              from: { type: string }
              to: { type: string }
              before_hash: { type: string }
              after_hash: { type: string }
              props: { type: object }
              changed:
                type: object
                additionalProperties:
                  type: object
                  properties:
                    before: {}
                    after: {}
              source_event_ids: { type: array, items: { $ref: "../schemas/common.v1.json#/$defs/uuid" } }
              attributed_to_event: { type: boolean }
    GraphLag:
      type: object
      required: [pending_jobs, in_flight_jobs, failing_jobs, oldest_unfinished_seconds, latency_p50_seconds, latency_p95_seconds, last_projected_at]
      properties:
        pending_jobs: { type: integer }
        in_flight_jobs: { type: integer }
        failing_jobs: { type: integer }
        oldest_unfinished_seconds: { type: number }
        latency_p50_seconds: { type: number }
        latency_p95_seconds: { type: number }
        last_projected_at: { type: [string, "null"], format: date-time }
...